allowed origins for pvt APIs, add public/private
