+++ API Server/apiserver.js
@@ -5,49 +5,36 @@
 
 const http = require('http');
 const url = require('url');
+const users = require('./publicAPIs/users');
+const commands = require('./privateAPIs/commands');
 
 module.exports = function(db) {
-    // users API
-    const users = (req, res) => {
-        if (req.method === 'GET') {
-            db.query("SELECT COUNT(*) AS users FROM channels", (err, results) => {
-                if (err) {
-                    res.writeHead(500);
-                    res.end(`ERROR: ${err}`);
-                    return;
-                }
-                const query = url.parse(req.url, true).query;
-                res.writeHead(200);
-                if (query.json !== undefined) {
-                    let responseObject = {
-                        'schemaVersion': 1,
-                        'label': 'users',
-                        'message': results[0].users.toString(),
-                        'color': 'blue'
-                    }
-                    res.end(JSON.stringify(responseObject));
-                } else {
-                    res.end(results[0].users.toString());
-                }
-            });
-        } else {
-            res.writeHead(400);
-            res.end('Bad Request');
-        }
-    }
-
     // API routes
     const apiRoutes = {
-        '/users': users,
+        public: {
+            '/users': users,
+        },
+        private: {
+            '/commands': commands,
+        }
     }
 
     // request handler
     const apiRequestHandler = (req, res) => {
+        const origin = req.ip || req.connection.remoteAddress || req.socket.remoteAddress || req.connection.socket.remoteAddress;
         const path = url.parse(req.url).pathname;
-        const handler = apiRoutes[path];
+
+        const allowedOrigins = process.env.ALLOWED_ORIGINS.split(',');
+        const isPrivateRequest = Object.keys(apiRoutes.private).includes(path);
+        if (isPrivateRequest && !allowedOrigins.includes(origin)) {
+            res.writeHead(401);
+            res.end('Unauthorized request to private API');
+        }
+        
+        const handler = isPrivateRequest ? apiRoutes.private[path] : apiRoutes.public[path];
         if (handler) {
             res.setHeader('Access-Control-Allow-Origin', '*')
-            handler(req, res);
+            handler(db, req, res);
         } else {
             res.writeHead(404);
             res.end('Not Found');
@@ -56,5 +43,5 @@ module.exports = function(db) {
 
     // basic http server
     const server = http.createServer(apiRequestHandler);
-    server.listen(process.env.PORT || 8080);
+    server.listen(process.env.PORT || 8080, '0.0.0.0', _ => {});
 }
