+++ API Server/privateAPIs/commands.js
@@ -1,0 +1,26 @@
+// Copyright (c) 2020 Mitchell Adair
+// 
+// This software is released under the MIT License.
+// https://opensource.org/licenses/MIT
+
+const url = require('url');
+
+const get = (db, req, res) => {
+
+}
+
+const post = (db, req, res) => {
+
+}
+
+const update = (db, req, res) => {
+
+}
+
+const remove = (db, req, res) => {
+
+}
+
+module.exports = (db, req, res) => {
+    
+}
